requirements.txt
asgiref>=3.8.1
Django>=5.1.3
django-browser-reload>=1.12.1
python-decouple>=3.8
sqlparse>=0.5.0
django-reversion>=5.1.0
django-crispy-forms>=2.3
crispy-bootstrap5>=2024.10
django-autocomplete-light>=3.11.0
bootstrap5>=24.3
django-bootstrap-datepicker-plus>=5.0.5
psycopg2>=2.9.10
django-reversion